main.js разбит на модули
