+++ src/utils.js
@@ -1,0 +1,6 @@
+// Render a template in certain block
+const render = (container, template, place) => {
+  container.insertAdjacentHTML(place, template);
+};
+
+export {render};
