+++ src/view/film-number.js
@@ -1,0 +1,5 @@
+export const createFilmNumberTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
